+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNGMV8YJAKWWNAM8W5SXRF6B"
+        "01KNK61YKA4FJ199BT3A0TBB2C"
     )
     r"""The unique identifier of the data source"""