+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNGMV8YXCGYKNA3C9GKY77GY"
+        "01KNK61YKPV6WJE8XWJHPW7SC7"
     )
     r"""The unique identifier of the data source"""