+++ packages/orq-rc/RELEASES.md
@@ -6869,3 +6869,13 @@ Based on:
 - [python v4.7.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.47 - packages/orq-rc
+
+## 2026-04-07 05:21:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.48 - packages/orq-rc